+++ internal/models/result.go
@@ -1,9 +1,9 @@
 ﻿package models
 
 type Result struct {
-    Subject string `json:"subject"`
-    Grade   string `json:"grade"`
-    Points  int    `json:"points"`
+    Subject string  `json:"subject"`
+    Grade   string  `json:"grade"`
+    Points  float64 `json:"points"`
 }
 
 type StudentResults struct {
